[Refactor] rename folder components to UI
