+++ src/layout/header/index.tsx
@@ -4,8 +4,8 @@ import { useMediaQuery } from "react-responsive";
 import { StyledMobileHeader, StyledDesktopHeader } from "./Header.styled";
 import { sizeBoundary } from "src/config/device";
 import { ReactComponent as Logo } from "src/assets/images/logo.svg";
-import { Button } from "src/components/MUI";
-import { AccountCircle, ArrowBack } from "src/components/MUI/icons";
+import { Button } from "src/UI/MUI";
+import { AccountCircle, ArrowBack } from "src/UI/MUI/icons";
 
 function MobileHeader() {
   return (
